Moved Category package inside core
