--- Tribler/Category/__init__.py
@@ -1,6 +1,0 @@
-# Written by Yuan Yuan
-# see LICENSE.txt for license information
-'''
-The category package contains code to categorize torrents and filter the 
-categories according to filter preferences.
-'''
