+++ Tribler/Core/Category/__init__.py
@@ -1,0 +1,3 @@
+"""
+The category package contains code to categorize torrents and filter the categories according to filter preferences.
+"""
