+++ Tribler/Core/Category/category.conf
