+++ Tribler/Core/Category/init_category.py
