+++ Tribler/Test/Core/Category/data/Tribler/Core/Category/category.conf
