+++ Tribler/Test/Core/Category/test_category.py
@@ -1,7 +1,9 @@
 import os
+
 from nose.tools import raises
-from Tribler.Category.Category import Category, cmp_rank
-from Tribler.Test.test_as_server import BaseTestCase, AbstractServer
+
+from Tribler.Core.Category.Category import Category, cmp_rank
+from Tribler.Test.test_as_server import AbstractServer
 
 
 class TriblerCategoryTest(AbstractServer):
@@ -55,7 +57,7 @@ class TriblerCategoryTest(AbstractServer):
         self.assertEquals(cmp_rank({'rank': 3}, {'bla': 4}), -1)
 
     def test_non_existent_conf_file(self):
-        import Tribler.Category.Category as category_file
+        import Tribler.Core.Category.Category as category_file
         category_file.CATEGORY_CONFIG_FILE = "thisfiledoesnotexist.conf"
         test_category = Category()
         self.assertEqual(test_category.category_info, [])
