+++ Tribler/Test/Core/Category/test_init_category.py
@@ -1,12 +1,13 @@
 import os
-from Tribler.Category.init_category import INIT_FUNC_DICT, getCategoryInfo
+
+from Tribler.Core.Category.init_category import INIT_FUNC_DICT, getCategoryInfo
 from Tribler.Test.test_as_server import AbstractServer
 
 
 class TriblerCategoryTestInit(AbstractServer):
 
     FILE_DIR = os.path.abspath(os.path.dirname(os.path.realpath(__file__)))
-    CATEGORY_TEST_DATA_DIR = os.path.abspath(os.path.join(FILE_DIR, u"data/Tribler/Category/"))
+    CATEGORY_TEST_DATA_DIR = os.path.abspath(os.path.join(FILE_DIR, "data", "Tribler", "Core", "Category"))
 
     def test_split_list(self):
         string = "foo ,bar,  moo  "
